Fix registration of embedded fonts
This is all a bit messy, but the gist is that there's no need to transfer fonts, or give them IDs, when they're eagerly registered during startup anyway.

Note, though, that they wouldn't necessarily have to be registered eagerly: we could also just build a list of embedded fonts' names and use that to find the correct symbols for HTML markup and `Font.defineFonts()`.